(feat) Thanks @felix @labjs, code now inserts base64 encoded speech into datastore, available for download from prompt.
Next steps are to a) read a dialog (or two) myself to guage how long it actually takes to read aloud, b) look at the file size
and determine if we need to gzip the recording too (hash it as well), c) start adding in other components, including assigning randomized affect and choice of handoff as well as measures. The indirect measure will be the most difficult.